requirements.txt
gunicorn
Flask
urllib3==1.24.2  # requests 2.21.0 has requirement urllib3<1.25,>=1.21.1
python-dotenv
dnspython  # sleekxmpp optional dependency
sleekxmpp
